Create concept of stateful peer.
This adds a `StatefulPeer` class that will be responsible for keeping
track of each peer connection.

I've also dropped all uses of "client", instead favoring the word
"peer". I've done this because "client code" is going to mean JS code
pretty soon, and I don't want that word to be overloaded.

Test Plan:
Test Plan:
Start the server (ie: `cargo run`) in one shell, then in another
connect to the server with telnet (ie: `telnet 127.0.0.1 13265`). Send
some text to the server and see it echo it to the console.